Создал базовый repository и service
