+++ api/src/domain/base/__init__.py
@@ -1,5 +1,9 @@
 from .entity import Entity
+from .repository import BaseRepository
+from .service import BaseService
 
 __all__ = [
     "Entity",
+    "BaseService",
+    "BaseRepository",
 ]
