+++ api/src/domain/base/repository.py
@@ -1,0 +1,23 @@
+from abc import ABC, abstractmethod
+
+
+class BaseRepository(ABC):
+    """
+    Базовый репозиторий
+    """
+
+    @abstractmethod
+    def get_by_id(self, pk: int):
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_list(self):
+        raise NotImplementedError
+
+    @abstractmethod
+    def delete(self, pk: int) -> None:
+        raise NotImplementedError
+
+    @abstractmethod
+    def exists(self, pk: int) -> bool:
+        raise NotImplementedError
